Update GUIListPanel.cpp
capturing the mouse to the list panel is redundant, since it is already captured to the list panel.
releasing the mouse leads to unexpected behavior, which is now fixed.
unexpected behavior: opening a list panel, then clicking a scrollbar, then clicking outside the list panel does not close the list panel. expected: close the list panel.